dealt with points having multiple candidate tracks in a semi-sensible way.
If a point wants to be in two edges, merge the track objects on
the fly.  If this happens in finger edge detection this is likely due
to a funny interaction with memory.  Rising/falling edges can not
actually cross or merge, so this should be OK in this case.  If this
is doing awful things, the search range is too big.